# Leverage `torch.addmm()` to speed up Newton-Schulz computation