Summary:
[`torch.addmm()`](https://docs.pytorch.org/docs/stable/generated/torch.addmm.html) fused the the matrix multiplication following and the addition of one matrix into kernel which speeds up the computation in Newton-Schulz significantly.

Use the following code to benchmark this:
```python
import torch
import time

def benchmark_matrix_ops(
    dim: int = 4096,
    iterations: int = 100,
    device: torch.device = None,
):
    if device is None:
        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")

    # Set random seed for reproducibility
    torch.manual_seed(42)
    A = torch.randn(dim, dim, device=device)
    X = torch.randn(dim, dim, device=device)

    # Scalars
    a, b, c = 1.1, 2.2, 3.3

    # Benchmark "Before" implementation
    start = time.time()
    for _ in range(iterations):
        B = b * A + c * A @ A
        X = a * X + X @ B
    time_before = (time.time() - start) / iterations

    # Re-initialize X for fair comparison
    A = torch.randn(dim, dim, device=device)
    X = torch.randn(dim, dim, device=device)

    # Benchmark "After" implementation using torch.addmm
    start = time.time()
    for _ in range(iterations):
        B = torch.addmm(A, A, A, beta=b, alpha=c)
        X = torch.addmm(X, X, B, beta=a, alpha=1.0)
    time_after = (time.time() - start) / iterations

    print(f"Device: {device}")
    print(f"Matrix dimension: {dim}x{dim}")
    print(f"Iterations: {iterations}")
    print(f"Before (manual ops) average time: {time_before * 1000:.3f} ms")
    print(f"After (torch.addmm) average time: {time_after * 1000:.3f} ms")

    return time_before, time_after

if __name__ == "__main__":
    benchmark_matrix_ops()
```

Averagely it shows 2x-5x speedup in 4096x4096 matrix for performing the benchmark ten times:
```bash
Device: cuda
Matrix dimension: 4096x4096
Iterations: 100
Before (manual ops) average time: 1.249 ms
After (torch.addmm) average time: 0.584 ms
Device: cuda
Matrix dimension: 4096x4096
Iterations: 100
Before (manual ops) average time: 1.220 ms
After (torch.addmm) average time: 0.583 ms
Device: cuda
Matrix dimension: 4096x4096
Iterations: 100
Before (manual ops) average time: 1.242 ms
After (torch.addmm) average time: 0.580 ms
Device: cuda
Matrix dimension: 4096x4096
Iterations: 100
Before (manual ops) average time: 1.247 ms
After (torch.addmm) average time: 0.584 ms
Device: cuda
Matrix dimension: 4096x4096
Iterations: 100
Before (manual ops) average time: 2.502 ms
After (torch.addmm) average time: 0.558 ms
Device: cuda
Matrix dimension: 4096x4096
Iterations: 100
Before (manual ops) average time: 1.218 ms
After (torch.addmm) average time: 0.585 ms
Device: cuda
Matrix dimension: 4096x4096
Iterations: 100
Before (manual ops) average time: 1.179 ms
After (torch.addmm) average time: 0.584 ms
Device: cuda
Matrix dimension: 4096x4096
Iterations: 100
Before (manual ops) average time: 2.245 ms
After (torch.addmm) average time: 0.556 ms
Device: cuda
Matrix dimension: 4096x4096
Iterations: 100
Before (manual ops) average time: 1.160 ms
After (torch.addmm) average time: 0.584 ms
Device: cuda
Matrix dimension: 4096x4096
Iterations: 100
Before (manual ops) average time: 2.854 ms
After (torch.addmm) average time: 0.563 ms
```

Differential Revision: D81389231